% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{abfith0}
\alias{abfith0}
\title{Beta parameters fit under the null}
\usage{
abfith0(dfH0)
}
\arguments{
\item{dfH0}{a dataframe with two columns 'a' and 'n', containing the 'a' samples from qmatchH0 with the corresponding n.}
}
\value{
list of estimated parameters (posterior means)
}
\description{
Beta parameters fit under the null
}
