% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{abfitha}
\alias{abfitha}
\title{Beta parameters fit under the alternative}
\usage{
abfitha(dfHA)
}
\arguments{
\item{dfHA}{a dataframe with two columns 'a', 'n', and 'theta' containing the 'a' samples from qmatchHA with the corresponding n and theta>0.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
list of estimated parameters (posterior means)
}
\description{
Beta parameters fit under the alternative
}
