% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSampDist-package.R
\docType{package}
\name{BayesSampDist-package}
\alias{BayesSampDist-package}
\alias{BayesSampDist}
\title{The 'BayesSampDist' package.}
\description{
estimatimates the sampling distribution of a Bayesian test statistic (posterior probability of alternative)
}
\references{
Stan Development Team (2022). RStan: the R interface to Stan. R package version 2.21.5. https://mc-stan.org
}
