% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pEst0}
\alias{pEst0}
\title{t1e estimate}
\usage{
pEst0(n, u, pars)
}
\arguments{
\item{n}{sample size.}

\item{u}{decision threshold.}

\item{pars}{list of beta parameter estimates under the null (alpha1, alpha2, sig0)}
}
\value{
vector of posterior mean and 95\% CIs for t1e
}
\description{
t1e estimate
}
