% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pEstA}
\alias{pEstA}
\title{power estimate}
\usage{
pEstA(n, theta, u, pars)
}
\arguments{
\item{n}{sample size.}

\item{theta}{effect size (linear scale).}

\item{u}{decision threshold.}

\item{pars}{list of beta parameter estimates under the alternative (beta1, beta2, sig1)}
}
\value{
vector of posterior mean and 95\% CIs for power
}
\description{
power estimate
}
