% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{power}
\alias{power}
\title{power fit and prediction main function}
\usage{
power(trcases, dat)
}
\arguments{
\item{trcases}{training set: sample and effect sizes at which simulations are available.}

\item{dat}{dataframe of simulated sampling distribution in training set.}

\item{testcases}{test/prediction set: a data frame of sample and effect sizes.}

\item{u}{decision threshold.}
}
\value{
dataframe of posterior mean and 95\% CIs for power in test set
}
\description{
power fit and prediction main function
}
