% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pred0}
\alias{pred0}
\title{t1e pred}
\usage{
pred0(testcases, u, pars)
}
\arguments{
\item{testcases}{test/prediction set: a vector of sample sizes.}

\item{u}{decision threshold.}

\item{pars}{list of beta parameter estimates under the null (alpha1, alpha2, sig0)}
}
\value{
data frame of posterior mean and 95\% CIs for t1e in test set
}
\description{
t1e pred
}
