% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predA}
\alias{predA}
\title{power pred}
\usage{
predA(testcases, u, pars)
}
\arguments{
\item{testcases}{test/prediction set: a data frame of sample and effect sizes.}

\item{u}{decision threshold.}

\item{pars}{list of beta parameter estimates under the alternative (beta1, beta2, sig1)}
}
\value{
data frame of posterior mean and 95\% CIs for power in test set
}
\description{
power pred
}
