% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qmatchHA}
\alias{qmatchHA}
\title{Quantile matching under the alternative}
\usage{
qmatchHA(
  ps,
  q = c(0.5, 0.6, 0.7, 0.8, 0.9, 0.95, 0.975, 0.98, 0.985, 0.99),
  ...
)
}
\arguments{
\item{ps}{Numeric vector of posterior probabilities (empirical distribution of the test-statistic).}

\item{q}{Numeric vector of tail probabilities corresponsing to the quantiles to be matched.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
A sample of shape parameter values for a beta fit with shape=1/scale parameters
}
\description{
Quantile matching under the alternative
}
